+++ sandbox_town_backend/src/main/java/com/shuidun/sandbox_town_backend/enumeration/Constants.java
@@ -2,5 +2,5 @@ package com.shuidun.sandbox_town_backend.enumeration;
 
 public class Constants {
     // 多少经验升一级
-    public static final long EXP_PER_LEVEL = 100;
+    public static final int EXP_PER_LEVEL = 100;
 }
